requirements.txt
appdirs==1.4.3
attrs==19.1.0
backcall==0.1.0
black==19.3b0
Click==7.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.3
django-braces==1.13.0
django-debug-toolbar==2.0
django-extensions==2.2.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.10.0
django-widget-tweaks==1.4.5
gunicorn==19.9.0
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.2
wcwidth==0.1.7
whitenoise==4.1.3
